+++ maintainer/download-data.sh
@@ -14,6 +14,7 @@ LHAPDF_SETS=(
   NNPDF40_nnlo_as_01175
   NNPDF40_nnlo_as_01185
   NNPDF40_nnlo_as_01190
+  ABMP16als118_5_nnlo
 )
 
 NEOPDF_SETS=(